# api/openapi.yaml
openapi: '3.1.0'
info:
  title: apikit API
  version: 0.1.0
  description: >-
    The apikit API provides user management, organization management,
    authentication, and authorization services.
servers:
  - url: /api/v1
    description: Default mount point (configurable per deployment)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Bearer token authentication supporting three credential types:
        Admin Token (format: <prefix>_admin_<64 hex chars>) for break-glass
        emergency access with global scope;
        API Key (format: <prefix>_<key_id>_<secret>) for user-scoped full
        access to own resources;
        PAT (format: <prefix>_pat_<token_id>_<secret>) for fine-grained
        per-permission access.
  headers:
    XRequestID:
      description: UUID for distributed tracing, present on every response
      schema:
        type: string
        format: uuid
...
    CacheControlNoStore:
      description: Cache-Control directive for mutable resources
      schema:
        type: string
        enum:
          - 'no-store'
...
  schemas:
...
    Error:
      type: object
      description: >-
        Standard error envelope. The code field equals the HTTP response
        status code exactly.
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: integer
              description: >-
                The HTTP status code of the response. Mirrors the HTTP status
                exactly — no additional error codes are used.
            message:
              type: string
              description: A human-readable error description.
...
    ApiKeyMetadata:
      type: object
      description: API key metadata returned in listings (no plaintext secret).
      required:
        - key_id
        - created_at
        - expires_at
        - revoked_at
      properties:
        key_id:
          type: string
          description: The random 8-character alphanumeric identifier portion of the API key.
        created_at:
          type: string
          format: date-time
          description: Timestamp when the API key was created (RFC 3339, UTC with Z suffix).
        expires_at:
          type:
            - string
            - 'null'
          format: date-time
          description: >-
            When the API key expires (RFC 3339 UTC). Null when expires=0
            (permanent, no expiry).
        revoked_at:
          type:
            - string
            - 'null'
          format: date-time
          description: Timestamp when the API key was revoked (RFC 3339, UTC). Null if not revoked.
    ApiKey:
      type: object
      description: API key object returned at creation or refresh, including the plaintext secret.
      required:
        - key
        - key_id
        - expires_at
      properties:
        key:
          type: string
          description: >-
            Full API key credential string including the plaintext secret.
            Format: <prefix>_<key_id>_<secret>. Only returned at creation or
            refresh — store securely.
        key_id:
          type: string
          description: The random 8-character alphanumeric identifier portion of the API key.
        expires_at:
          type:
            - string
            - 'null'
          format: date-time
          description: >-
            When the API key expires (RFC 3339 UTC). Null when expires=0
            (permanent, no expiry).
...
  responses:
...
    Unauthorized:
      description: Unauthorized — missing, invalid, expired, or revoked credentials.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden — valid credentials but insufficient permissions or user blocked.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found — the requested resource does not exist.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /user/keys:
    get:
      operationId: getUserKeys
      summary: List user's API keys
      description: >-
        List the authenticated user's API key(s). Accepts Admin Token,
        API Key, or PAT with keys:read permission.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of API key metadata.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKeyMetadata'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /user/keys/{key_id}:
    delete:
      operationId: deleteUserKey
      summary: Revoke user's API key
      description: >-
        Revoke the authenticated user's own API key. Accepts Admin Token,
        API Key, or PAT with keys:manage permission.
      security:
        - bearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
          description: The API key identifier.
      responses:
        '204':
          description: API key revoked successfully. No body.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/keys/{key_id}/refresh:
    post:
      operationId: refreshUserKey
      summary: Refresh user's API key
      description: >-
        Refresh the authenticated user's API key with a new secret. Accepts
        Admin Token or API Key only; PAT credentials are rejected with 401
        regardless of their permissions.
      security:
        - bearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
          description: The API key identifier.
      responses:
        '200':
          description: Refreshed API key with new secret.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...